+++ src/base/VolumeInterface.php
@@ -171,8 +171,27 @@ interface VolumeInterface extends SavableComponentInterface
      * @param string $path The path of the directory, relative to the source’s root
      * @throws VolumeObjectExistsException if a directory with such name already exists
      * @throws VolumeException if something else goes wrong
+     * @deprecated in 3.6.0. Use [[createDirectory()]] instead.
      */
     public function createDir(string $path);
+
+    /**
+     * Creates a directory.
+     *
+     * @param string $path The path of the directory, relative to the source’s root
+     * @throws VolumeObjectExistsException if a directory with such name already exists
+     * @throws VolumeException if something else goes wrong
+     */
+    public function createDirectory(string $path);
+
+    /**
+     * Deletes a directory.
+     *
+     * @param string $path The path of the directory, relative to the source’s root
+     * @throws VolumeException if something goes wrong
+     * @deprecated in 3.6.0. Use [[deleteDirectory()]] instead.
+     */
+    public function deleteDir(string $path);
 
     /**
      * Deletes a directory.
@@ -180,7 +199,19 @@ interface VolumeInterface extends SavableComponentInterface
      * @param string $path The path of the directory, relative to the source’s root
      * @throws VolumeException if something goes wrong
      */
-    public function deleteDir(string $path);
+    public function deleteDirectory(string $path);
+
+    /**
+     * Renames a directory.
+     *
+     * @param string $path The path of the directory, relative to the source’s root
+     * @param string $newName The new path of the directory, relative to the source’s root
+     * @throws VolumeObjectNotFoundException if a directory with such name already exists
+     * @throws VolumeObjectExistsException if a directory with such name already exists
+     * @throws VolumeException if something else goes wrong
+     * @deprecated in 3.6.0. Use [[renameDirectory()]] instead.
+     */
+    public function renameDir(string $path, string $newName);
 
     /**
      * Renames a directory.
@@ -191,5 +222,5 @@ interface VolumeInterface extends SavableComponentInterface
      * @throws VolumeObjectExistsException if a directory with such name already exists
      * @throws VolumeException if something else goes wrong
      */
-    public function renameDir(string $path, string $newName);
+    public function renameDirectory(string $path, string $newName);
 }