+++ CHANGELOG.md
@@ -8,8 +8,11 @@
 - Added the `handleCasing` config setting, which determines the default casing that should be used when autogenerating component handles. ([#4276](https://github.com/craftcms/cms/issues/4276))
 - Added `craft\base\ApplicationTrait::getFormattingLocale()`, which returns the locale that should be used for date/time formatting.
 - Added `craft\base\FieldInterface::useFieldset()`, which custom fields can override to return `true` if a `<fieldset>` and `<legend>` should be used, rather than a `<div>` and `<label>`.
+- Added `craft\base\VolumeInterface::createDirectory()`.
+- Added `craft\base\VolumeInterface::deleteDirectory()`.
 - Added `craft\base\VolumeInterface::getDateModified()`.
 - Added `craft\base\VolumeInterface::getFileSize()`.
+- Added `craft\base\VolumeInterface::renameDirectory()`.
 - Added `craft\fieldlayoutelements\BaseField::useFieldset()`.
 - Added `craft\fields\Url::TYPE_EMAIL`.
 - Added `craft\fields\Url::TYPE_TEL`.
@@ -35,7 +38,10 @@
 - Updated Yii to 2.0.40.
 
 ### Deprecated
+- Deprecated `craft\base\VolumeInterface::createDir()`. `createDirectory()` should be used instead.
+- Deprecated `craft\base\VolumeInterface::deleteDir()`. `deleteDirectory()` should be used instead.
 - Deprecated `craft\base\VolumeInterface::getFileMetadata()`. `getFileSize()` and `getDateModified()` should be used instead.
+- Deprecated `craft\base\VolumeInterface::renameDir()`. `renameDirectory()` should be used instead.
 - Deprecated `craft\helpers\Template::paginateCriteria()`. `paginateQuery()` should be used instead.
 
 ### Removed